# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS AI - RPG-style AI skill system API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AiProfile {
  id: number;
  name: string;
  level: number;
  xp: number;
  xpToNext: number;
  aiClass: string;
  description: string;
  personality: string;
  createdAt: string;
}

export interface AiProfileUpdate {
  name?: string;
  personality?: string;
}

export interface AiStats {
  totalSkills: number;
  acquiredSkills: number;
  pendingSkills: number;
  fusedSkills: number;
  totalXp: number;
  level: number;
  recentActivity: string[];
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  title: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface Skill {
  id: number;
  name: string;
  category: string;
  description: string;
  level: number;
  /** acquired | pending | fused */
  status: string;
  xpValue: number;
  icon: string;
  color: string;
  principles: string[];
  /** @nullable */
  parentSkillIds?: number[] | null;
  /** @nullable */
  studySource?: string | null;
  createdAt: string;
}

export interface SkillInput {
  name: string;
  category: string;
  description: string;
  icon?: string;
  color?: string;
}

export interface SkillUpdate {
  name?: string;
  description?: string;
  status?: string;
}

export interface StudyRequest {
  /** Topic, URL, or repository to study */
  topic: string;
  /**
   * What to include/exclude from study
   * @nullable
   */
  filter?: string | null;
}

export interface StudyResult {
  summary: string;
  proposedSkills: Skill[];
}

export interface FuseRequest {
  /** @minItems 2 */
  skillIds: number[];
  /** @nullable */
  fusionName?: string | null;
}

export interface FuseResult {
  fusedSkill: Skill;
  description: string;
}

export interface SkillCategory {
  name: string;
  count: number;
  icon: string;
  color: string;
}

export type ListSkillsParams = {
  category?: string;
  status?: string;
};