# lib/api-zod/src/generated/types/aiProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS AI - RPG-style AI skill system API
 * OpenAPI spec version: 0.1.0
 */

export interface AiProfile {
  id: number;
  name: string;
  level: number;
  xp: number;
  xpToNext: number;
  aiClass: string;
  description: string;
  personality: string;
  createdAt: Date;
}