# lib/api-zod/src/generated/types/aiProfileUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS AI - RPG-style AI skill system API
 * OpenAPI spec version: 0.1.0
 */

export interface AiProfileUpdate {
  name?: string;
  personality?: string;
}