# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS AI - RPG-style AI skill system API
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: number;
  name: string;
  category: string;
  description: string;
  level: number;
  /** acquired | pending | fused */
  status: string;
  xpValue: number;
  icon: string;
  color: string;
  principles: string[];
  /** @nullable */
  parentSkillIds?: number[] | null;
  /** @nullable */
  studySource?: string | null;
  createdAt: Date;
}