# lib/api-zod/src/generated/types/studyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS AI - RPG-style AI skill system API
 * OpenAPI spec version: 0.1.0
 */

export interface StudyRequest {
  /** Topic, URL, or repository to study */
  topic: string;
  /**
   * What to include/exclude from study
   * @nullable
   */
  filter?: string | null;
}